requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.3.0
requests==2.32.3
PyJWT==2.8.0
python-decouple==3.8
mysqlclient==2.2.4